Relocate output functions
