+++ libs/compiler/frontend_utils.h
@@ -16,6 +16,8 @@
 
 #pragma once
 
+#include "context.h"
+
 
 #ifdef __cplusplus
 extern "C" {
@@ -27,30 +29,6 @@ extern "C" {
  *	@param	context	RuC context
  */
 void read_keywords(compiler_context *context);
-
-/**
- *	Print tables and tree to a file specified by @p path
- *
- *	@param	context	RuC context
- *	@param	path	Target file
- */
-void output_tables_and_tree(compiler_context *context, const char *path);
-
-/**
- *	Print codes to a file specified by @p path
- *
- *	@param	context	RuC context
- *	@param	path	Target file
- */
-void output_codes(compiler_context *context, const char *path);
-
-/**
- *	Print export tables to a file specified by @p path
- *
- *	@param	context	RuC context
- *	@param	path	Target file
- */
-void output_export(compiler_context *context, const char *path);
 
 #ifdef __cplusplus
 } /* extern "C" */
